Add add_simulation_steps method to model.py
It is annoying when one has run a simulation to the end, that one cannot say "ok that was interesting, but I would now like to continue that same simulation for another n steps", with the current implementation that was not possible. I added the add_simulation_steps method in order to add extra steps to the end-time of the simulation. And modified the rest so that the same random gen state would be kept as well as the state of the simulation, and after calling add_simulation_steps one can then simply call run again which will keep the states of the simulation and simply continue the simulation as if it had been called with those extra steps from the start.